requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
pymongo==4.9.2
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
pydantic[email]==2.12.3
python-multipart==0.0.20
pandas==2.3.3
numpy==2.3.4
scikit-learn
spacy==3.7.4
nltk==3.9.1
sklearn-crfsuite==0.3.6
